split base_tfrecord
