+++ src/datumaro/plugins/data_formats/roboflow/base.py
@@ -6,7 +6,6 @@ import csv
 import errno
 import os
 import os.path as osp
-import re
 from typing import Dict, List, Optional, Union
 
 from defusedxml import ElementTree
@@ -317,40 +316,3 @@ class RoboflowMulticlassBase(SubsetBase):
                 )
 
         return items
-
-
-try:
-    from datumaro.plugins.data_formats.tf_detection_api.base import TfDetectionApiBase
-    from datumaro.plugins.data_formats.tf_detection_api.format import TfrecordImporterType
-
-    class RoboflowTfrecordBase(TfDetectionApiBase):
-        def __init__(
-            self,
-            path: str,
-            *,
-            subset: Optional[str] = None,
-            ctx: Optional[ImportContext] = None,
-        ):
-            super().__init__(
-                path=path,
-                subset=subset,
-                tfrecord_importer_type=TfrecordImporterType.roboflow,
-                ctx=ctx,
-            )
-
-        @staticmethod
-        def _parse_labelmap(text):
-            entry_pattern = r'name:\s*"([^"]+)"\s*,\s*id:\s*(\d+)'
-            entry_pattern = re.compile(entry_pattern)
-
-            matches = re.findall(entry_pattern, text)
-
-            labelmap = {name: int(id) for name, id in matches}
-
-            return labelmap
-
-except ImportError:
-
-    class RoboflowTfrecordBase:
-        def __init__(self):
-            raise ImportError("Tensorflow package not found. Cannot convert to Tensorflow dataset.")
